+++ CHANGELOG.md
@@ -9,4 +9,5 @@ All notable changes to this project will be documented in this file.
 - Depend on OxyPlot.Core 1.0.0
 - Build OxyPlot.GtkSharp with GtkSharp 2.12.45
 - Build OxyPlot.GtkSharp3 with GtkSharp 2.99.3
+- Mitigate memory issue with Pango.Layout #11
 - Fix application freeze on Linux/X11 when viewing "show tracker without clicking" example (#13)